Initial commit
Simple version with connection between arduino-PC-server

Arduino firmware can accept simple handshake arguments from the node
server to turn on/off some digital channels.

The "server" connects to the arduino and sends the data (via websocket)
to the connected clients.

Some clients (written in javascript) plot the data in "real-time" and
are able to interact with the arduino, turning on/off some of the
channels.