Changed node for uninterrupted serial connection
To accomodate the genuino Mega, I changed the code in the firmware, which (for some still-unexplained reason) took about 1 second to initialize the serial communications everytime the arduino had to connect.

Under the previous architecture this was unacceptable, since every message from the master server triggered a new serial connection- which was, in any case, quite suboptimal.

Now I have changed the node to initialize the serial connection upon starting the node, and just keep track and restart the connection when necessary.