Updated handshake function and fixed bugs
The most persisten bug appeared whenever I was reading from the UNO:
I was receiving a "Nonetype cannot be iterated" error after calling the
SerialCommManager.poll_arduino() in the node. The reason for this is
that if the received response for the arduino is "None", or if it did
not comply with the length in the NUM_CHANNELS variable in
communications.SerialCommManager, the function would not return
anything. However, the output of the function was being assigned to two
different variables (t, channels), and the result of this assignment was
receiving nothing.

Also, I have updated the documentation of the servers.server_master