Removal of multiple-data-input
During the polling of arduino, the (legacy) code could accept several
concatenated sets of values for the channels, with the assumption that
reading the data and sending it multiple times carried more overhead
than reading the data several times and sending a whole bunch at once.

Since we do not currently pre-process the data, but rather just obtain a
single value at each polling period, I have removed the code that
performed the slicing and reshaping of the data for clarity and
simplicity.

I have tested with both the UNO code and the MEGA