Added lab metadata notification
Maybe not the best coding of my life, but now, thanks to the observer
pattern, I was able to append metadata to the database entries.

A function in the master called "db_metadata_append" takes care of
appending the metadata to the database. The master does so by observing
the value of the comms_handler.last_metadata_id value, which changes
when the Nodes send a configuration dictionary.
Such a configuration dictionary is characterised by having the key
'conf'.

NOTE: new nodes should have an updated server_node.DICT_CONTENTS
variable!