Fixed some bugs on registering the nodes
Now nodes are identified also with the "user" identification (lab6,
lab7,...) in the status page.
Since some errors were cropping up (the user not being registered to the
master server all the time because the "conf" dictionary was not being
sent upon reconnection), decided to set the metadata_registered property
to False in a bunch of error cases to make sure this configuration
dictionary is sent after each reconnection.
Also added some cosmetic changes to the status page (the labs appear
horizontally distributed)