Added NODES.md and document bug
Added more in-depth documentation about the nodes.

Also, made explicit an issue where a new node would fail to register
because the metadata dictionary did not contain the keys 'user' or
'error'.

This same problem would prevent the master server from storing data from
a node if the dictionaries corresponding to the data contained keys
different from those of the metadata.